requirements.txt
# Core framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.12.1

# AI/ML Core
openai==1.6.1
faster-whisper==1.0.0
pyannote.audio==3.3.2
transformers==4.35.2
torch==2.2.0
torchaudio==2.2.0

# NLP and Analysis
langchain==0.0.335
langchain-openai==0.0.2
umap-learn==0.5.4
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
nltk==3.8.1

# Visualization and UI
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Web and API
requests==2.31.0
httpx==0.25.2
websockets==12.0

# Utils
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
tenacity==8.2.3
typing-extensions==4.8.0

# Development
pytest==7.4.3
black==23.11.0